Asset folder and a Zip file
Hey guy! I have started working on the project and this is what I have so far. The asset folder is probably what you need to copy to your file folder if you plan to work on it or just have a look. Else if that doesn't work, the zip file is the whole project I created and worked on since I can't work on Rachel files.

Thanks